+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD
 
+- CLI: tx update-config accepts json
+
 ## v0.4.4
 
 - fix: iovnscli get config